[FIX] l10n_mx_reports: Fixing Diot tags.
Description of the issue/feature this PR addresses:
The language used in the DIOT documentation was confusing and even though we had the exact same description for two columns it turns out they were different, so we need to change the logic of a few columns to make it work as needed.
This issue was addressed in PR: https://github.com/odoo/odoo/pull/217441, but a small adjustment still needs to be made in the l10n_mx_reports module.

Current behavior before PR:
The l10n_mx_diot_get_values function uses references to the old account.report.expression that were reworked in the DIOT columns in the pr mentioned above.

Desired behavior after PR is merged:
The l10n_mx_diot_get_values function uses the new references to account.report.expression that were introduced in the pr of the community repo.

opw-4920577

"I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr"

fix

closes odoo/enterprise#90585

X-original-commit: 2ccf1d8ca6ab5d9b7c181a580427e48e4322a79f
Related: odoo/odoo#219784
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>
Signed-off-by: Hector Mauricio Perez Barajas (hmpb) <hmpb@odoo.com>