[REF] *: cleanup datetime service API
This commit is a rework of the DateTime service & hook modules in web.

The API of the 'datetime_service' was too convoluted, bloated with
exported members that were only used by the hook.

Now, most of the logic is handled by the service, and the hook simply
gives it a flag to tell the service to handle Owl lifecycle hooks. This
effectively allows to only send the few members (methods & variables)
that are used by callers of the service externally.

Task 3495251

closes odoo/enterprise#89299

Related: odoo/odoo#217179
Signed-off-by: Bastien Fafchamps (bafa) <bafa@odoo.com>