[FIX] sale_sub: pricing copy on template copy
If a pricelist rule was restricted to a specific variant, even
if the template doesn't have multiple variants anymore, it should
still be copied for the matching variant.

If a rule was targeting an inactive product, duplication failed
because no matching product was found on the new template.
Those rules can be safely ignored on duplication

X-original-commit: 61b91bb5748af9430eec1e13c6f0be26e8031a82
Part-of: odoo/enterprise#89574
Related: odoo/odoo#217598
Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>