[FIX] planning: prevent creation of shifts with no resource
This commit's purpose is to prevent the creation of open shifts in the
calendar view of the planning. Recently, the multi-create feature was
added to the calendar view, but some code of the early iteration of the
task was left. This fix aim to remove the unneeded code.

target saas-18.4

closes odoo/enterprise#89565

X-original-commit: 85448a35f0c85c065bdc847281df42d3a8f6f6a5
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>