[FIX] l10n_be_reports: modify company check on partner to acommodate with auto_check
Steps to reproduce:

- Install l10n_be_reports & account_accountant.
- Switch to "BE Company CoA" company.
- Add "281.50" tag to "YourCompany" contact.
- Accounting > Configuration > Accounting > Chart of Accounts
  > Add "281.50 - Commissions" tag to 600000 account in CoA.
- Create a Vendor Bill for partner "YourCompany" selling "[E-COM11]".
  Date it sometime in 2024 and confirm.
- Go to Accounting > Reporting > Open 325 Forms.
- Create a new form for 2024.
- Verify that there is a line for "YourCompany" in the form and
  generate the XML.
- Try to install Inventory.
> Traceback due to a company inconsistencies.

Cause of the issue:

When creating the new `l10n_be.form.281.50` form for 2024, we create a
record for "BE Company CoA"  and which `partner_id` is the `res.partner`
"YourCompany" which does not belong to any company (and hence is
accessible from any). However, on the `forms_281_50` related field
of the `res.partner` model is `check_company=True`:
https://github.com/odoo/enterprise/blob/bc36a68b68c3d0296ee18c3a121b029dd418f430/l10n_be_reports/models/res_partner.py#L15-L23
Currently, this did not matter because we did no call of the
`_check_company`. But if we where to launch a call of the
`_check_company` on "YourCompany" `res.partner` we would raise a
company error. Now, when you install stock 2 things happen and provoke
this call:
1) The `res.partner` model becomes `_check_company_auto`:
https://github.com/odoo/odoo/blob/b7dbf15aa8116e51708bd51b8dda9bd12bba0a9e/addons/stock/models/res_partner.py#L8-L10
2) We create a warehouse for "Your company and update the `partner_data`
https://github.com/odoo/odoo/blob/b7dbf15aa8116e51708bd51b8dda9bd12bba0a9e/addons/stock/models/stock_warehouse.py#L143-L145
https://github.com/odoo/odoo/blob/b7dbf15aa8116e51708bd51b8dda9bd12bba0a9e/addons/stock/models/stock_warehouse.py#L315
However, the `property_stock_customer` and `property_stock_supplier`
are `check_company` related fields:
https://github.com/odoo/odoo/blob/b7dbf15aa8116e51708bd51b8dda9bd12bba0a9e/addons/stock/models/res_partner.py#L12-L17
Therefore, writting on them for a `_check_company_auto` record
will launch a call of the `check_company` for all the fields of the
record (including the already inconsistent `forms_281_50`):
https://github.com/odoo/odoo/blob/b7dbf15aa8116e51708bd51b8dda9bd12bba0a9e/odoo/models.py#L3780-L3781
https://github.com/odoo/odoo/blob/b7dbf15aa8116e51708bd51b8dda9bd12bba0a9e/odoo/models.py#L3867-L3868

Fix:

We loosen the company check on the `forms_281_50` to not rely on the
`_check_company` since the current workflow lead functionally to valid
record data's that will always be considered invalid by this method.

opw-4845434

closes odoo/enterprise#90088

X-original-commit: 5275f44810a980d77455e1a4a8e70c0d73c66f8f
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Lancelot Semal (lase) <lase@odoo.com>