[FIX] hr_contract_salary: generate new offer link
_______________________________________
When creating an offer on the page of another offer, the salary
configurator url doesn't contain a token, but instead, contains
token = false.

1. Open the recruitment module. If there's no job position nor
 application, create some.
2. Open an existing application and click on Generate Offer.
3. Click on New on the top left of the screen.
4. Fill the new application and click on save.

The field 'Link' contains an URL with, at the end, &token=x, with x a
randomly generated token.

The field 'Link' contains an URL with, at the end, &token=False.

The token generation code wasn't called when creating a new offer the way
described in reproduction steps.

_________________________________________
opw-4885796

---

closes odoo/enterprise#98593

X-original-commit: cce12e96ab823c59f7af133b9ad2fe18f0ff0968
Signed-off-by: Maeva de Keyser (mdek) <mdek@odoo.com>
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>