[FIX] voip: replace cross icon with a minus icon
The "cross icon" (X) is believed to be too "scary"; people don't dare to
click it in the middle of a call out of fear of hanging up, even though the
softphone window might be in their way.

The "minus icon" (-) might deceive one's expectations by hiding the
window rather than minimizing it, but we decided that it was still
better this way.

Part of task-4891947.

closes odoo/enterprise#89691

X-original-commit: fa59bb332f73b0fab50c88cb0ed6701e6459f71b
Signed-off-by: Yuchen Huang (yhu) <yhu@odoo.com>
Signed-off-by: Louis Wicket (wil) <wil@odoo.com>