[IMP] iot: avoid waiting for longpolling timeout when failure
To contact an IoT Box, if longpolling fails, we fallback to websocket.
We use this logic on every call: if the longpolling fails once, there is
chance that it fails on the next call.

We now avoid contacting via longpolling if it has failed during the last
20 minutes. We then avoid 6 seconds timeout on every call on networks
where local requests fail (clients with DNS for example).

closes odoo/enterprise#90441

Task: 4922624
X-original-commit: d836e56467bf3f93dabdd70bfd1d9993bb8f7275
Signed-off-by: Max Whale (mawh) <mawh@odoo.com>
Signed-off-by: Louis Travaux (lotr) <lotr@odoo.com>