[IMP] spreadsheet_edition: track real revision date in version history
Currently, when duplicating a spreadsheet along with its revisions,
each duplicated revision is assigned the `create_date` of the new
record, resulting in all revision timestamps showing the exact same
datetime. This prevents users from being able to trace the history
of changes and makes it difficult to identify when a specific revision
was originally made. To address this issue, a new field of `datetime`
type `revision_date` is introduced to stores the actual datetime when
the original revision occurred. The version history side panel now displays
this `revision_date` field instead of relying on `create_date`, enabling
users to accurately track the evolution of spreadsheet changes over time.
NOTE: For the 'original data' revision and in order to avoid adding
a new field on the spreadsheet model, the timestamp will be hidden.

Task-4854898

closes odoo/enterprise#89647

Related: odoo/upgrade#8023
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>