[IMP] l10n_*_hr_payroll: don't set password on users
These users are created in order to create demo / test
employees. They're not used for tours or other log-in requirements. As
such they don't need a password, and all that does is burn CPU (as the
password needs to be hashed on user creation). The cost is not huge in
the grand scheme of things, but it's unnecessary.

If someone wants to log out as a payroll demo employee for reasons,
they can set a password by hand. Same for tests.

closes odoo/enterprise#89614

Signed-off-by: Xavier Morel (xmo) <xmo@odoo.com>