[FIX] spreadsheet_edition: correctly save field matching
Steps to reproduce:
1. Open a spreadsheet with global filters.
2. Open the global filter editor.
3. Open a global filter.
4. Delete the field matching (and nothing else).
5. Save the global filter.
=> The field matching is not saved.

This was due to the fact that the removal of the field matching makes an
early return in the `updateFieldMatching` method, which had for last
instruction to set the `draft` property to mark the filter as dirty.

This commit fixes the issue by ensuring that the `draft` property is set
even when the field matching is removed.

closes odoo/enterprise#89793

Task: 4882261
X-original-commit: 106f93bfdda060a38daa2d06a26e8abe9696e794
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>