[FIX] sign: handle false validity in sign requests during reminders
When a 'sign.request' record has 'validity' as 'False', attempting to send a
reminder via the '_cron_reminder' method leads to a crash.

Steps to Reproduce:-
1. Install the 'Sign' module.
2. Navigate to the Sign section and click on 'Upload PDF & Sign.'
3. Upload any PDF document and add your signature, then click 'Send.'
4. In the new wizard, remove the value for 'Valid Until' and enable the
 'reminder' option. Set the reminder to 'every 1 day.'
5. When our scheduled action named 'Sign: Send Mail Reminder' executes
 the following day, it will throw an error.

Error:-
'TypeError(''<' not supported between instances of 'bool' and 'datetime.date'')
while evaluating 'model._cron_reminder()''

Root Cause:-
The SQL query within the '_cron_reminder()' method retrieves all records where:
- The request is active and in the sent state.
- Either:
  - 'validity < today' or
  - A reminder is due based on 'last_reminder + reminder'.

The fetched records are then iterated through at [1].
[1] https://github.com/odoo/enterprise/blob/ac4aeeea98dcf2fc7f06e6a3fabc55e256330e2c/sign/models/sign_request.py#L454

If 'validity' is 'False', this comparison raises a 'TypeError'
because it is invalid to compare a 'boolean' with 'datetime.date'.

Solution:-
- A safety check was added before the comparison between 'request.validity' and
  today's date, ensuring that 'request.validity' exists.

Sentry-6727599497

closes odoo/enterprise#91324

X-original-commit: 050ddbca6ad829c2d9bdd09997835d12cd2a12e6
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Krishna Arvindkumar Patel (krip) <krip@odoo.com>