[REF] web_tour: move step_utils
In this commit, we move step_utils to the root of web_tour/static/src.
The goal is to move all the files that are in tour_service into
web_tour/static/src and then split it into tour_automatic and
tour_interactif in order to create 2 separate bundles and be able
to lazy load what tour_automatic needs only when we launch an automatic
tour.

closes odoo/enterprise#89712

Related: odoo/odoo#217854
Related: odoo/design-themes#1111
Signed-off-by: Bastien Pierre (ipb) <ipb@odoo.com>