[FIX] mrp_workorder: create warehouse for tour
Instead of using default created warehouse, this commit creates a new
warehouse to use it in the tour.
The purpose of this change is to avoid name's conflict with MO created
by demo data and by the way, it's better to use specific to test's data.

Rnbot build error: 223328

X-original-commit: 609c9dc4bc4c1d4543db2fe1058ee440c7a49b29
Part-of: odoo/enterprise#89777
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>