[FIX] sale_commission: include last day of commission period
**Issue**
Achievements realised on the last day of the period were not taken
into account (neither in the expected one, or the next period).

**Steps to reproduce**
- Create a commission plan with a monthly target frequency and
an achievement on "Amount sold"
- Have a Sale order with a `date_order` on the last day of a
month
- Click on the "Commissions" smart button to open the
`sale.commission.report` list view -> achieved amount for the
SO's month is missing.
- Click on "Details" for the line -> there should be one
`sale.commission.achievement.report` record.

**Cause**
We were comparing a Timestamp coming from the commission line with
the start/end date of the period.
To perform the comparison, the date was cast to a Timestamp with
0:00:00 as the hour, resulting in all commission lines on the last
day of the period to be missed.

opw-4857936
opw-4934168

closes odoo/enterprise#91748

X-original-commit: 98178673806eabe032f519ea24deabe72b7c8449
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Tanguy Quéguineur (taqu) <taqu@odoo.com>