[IMP] l10n_au_*: salary in tax closing entry
module: l10n_au_hr_payroll_account

This commit is AU specific and will create the lines related to tax on
salary in the tax return closing entry. Before this commit, these lines
weren't taken into account in the closing entry.
The condition to select the account.move.line are the following:
- The journal item has a "W2" or "W3" (Withholding Tax) tax grid whose
journal entry is linked to an hr.payslip.
- The journal item does not have an originating tax record (i.e.,
tax_line_id is empty).
- The journal item is posted to a tax provision account (i.e., a balance
sheet one) as defined on the salary rule.

Task-4921992

X-original-commit: 65ea5fdfbf5d95b79b2fd1d09b2d881309db93ea
Part-of: odoo/enterprise#92044
Signed-off-by: William André (wan) <wan@odoo.com>