[IMP] pos_blackbox_be: remove reload drivers popup at install
When we installed the module `pos_blackbox_be`, a dialog was shown to ask the user to
reload his IoT Box's drivers (in order for its blackbox to be dected as a blackbox
instead of an adam scale).

We removed this dialog as we now download the blackbox driver by default on IoT Boxes.

closes odoo/enterprise#89899

X-original-commit: af860cba64c0afa10b384fb1105449f72baf39b6
Related: odoo/upgrade#8055
Signed-off-by: Max Whale (mawh) <mawh@odoo.com>
Signed-off-by: Louis Travaux (lotr) <lotr@odoo.com>