[IMP] appointment, *: make avatars_display a boolean field
*: {website_}appointment_{account_payment,hr_recruitment}

As the option is an on/off parameter, we move from a selection
field to a boolean one. It has a few advantages:
It makes the form a bit lighter, it saves space in the db, and it
makes the code more readable.

We adapt the various occurences of the parameter and update the
related snippet options. Indeed, the parameter is also linked to a
website editor option that toggles it.

The new field is called show_avatars to match the wording of the
snippet, and also because we use placeholder images when disabled
on the operator selection screen -> 'avatars' is more precise than
'pictures'.

UPG PR: odoo/upgrade#8014

Task-4898893

Part-of: odoo/enterprise#88637
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>