[FIX] delivery_{dhl,usps}_rest: fix getting packages details
This commit changes the method name for retrieving packages details when
shipping with USPS or DHL. Previously, the method used for that had the
same name in both connectors. That led to sending the wrong details
depending on which module got installed first, because they are the same
model `delivery.carrier`. This commit ensures both functionalities are
completely separate by changing the method signature.

closes odoo/enterprise#90383

X-original-commit: d86ff678a087aea674cc365982221f30a52da40b
Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>
Signed-off-by: Mohammad Abdulmoneim (abdu) <abdu@odoo.com>