[FIX] l10n_be_hr_contract_salary: ensure plan_to_change_car is set to False after signing a contract.
Issue: When a contract was signed and a future driver was assigned to a vehicle, the car still remained marked as available.
Fix: Once the first signature is completed and the future driver is assigned, the vehicle is correctly marked with plan_to_change_car = False (same for the bike).
Related task: 4926335.

closes odoo/enterprise#93307

X-original-commit: 65f1bff8cccffb8991721e0b1a573ecadd8c7aa3
Related: odoo/odoo#224463
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>