[FIX] l10n_uy_edi: Free Delivery line zero (invoice Indicator 5)
This PR fixes an issue where invoice lines with zero unit price were not
being reported in the CFE (Comprobante Fiscal Electrónico) for Uruguayan
electronic invoicing. The fix ensures that lines with zero price are
correctly classified as "Free Delivery" (invoice indicator 5) and
included in the CFE XML.

Before this changes if we have a line with price unit != 0.0 but with
total price of the line 0.0 then we are reporting the invoice line as
Free Delivery. But, If we have an invoice with line with price unit 0.0
then is not being informed in the CFE at all

After this changes both lines are informed to DGI using the invoice
indicator 5 (Free Delivery). Any line that totals equal 0.0 will be
informed as Free Delivery (Indicator 5)

References
LATAM 1350 / ADHOC task 53445

closes odoo/enterprise#97262

X-original-commit: c309e32a7f9a2219a0076ca213f6cc8a46bf61ef
Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>