[IMP] helpdesk: remove server action for customer ratings
Before this commit, the "Customer Ratings" action in helpdesk menuitem was a
server action. This was incompatible with the framework's offline mode, as it
required a live connection to the server for execution.

This commit replaces the server action with a direct window action. The dynamic
company filtering is now handled by a non-stored computed field with a search
method.

task-4920367

closes odoo/enterprise#90301

Related: odoo/upgrade#8162
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>