[IMP] appointment: add dynamic button for video link
Once a location is set, the video link is less used as it will
often be an onsite appointment, but can still be set up. In order
to make it available in that case, but hidden when location is set
by default, we add a small button that sets the videocall_source to
'discuss' when clicked, hence making the field appear. Emptying the
field will make it disappear and make the button reappear.

Task-4898893

Part-of: odoo/enterprise#88637
Related: odoo/upgrade#8014
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>