[IMP] industry_fsm: remove server action for customer ratings
Before this commit
The "Customer Ratings" action in industry_fsm used a server action, incompatible
with offline mode due to its dependency on a live backend connection.

After this commit
The server action is replaced with a standard window action, which is supported
in offline mode.

task-4920361

closes odoo/enterprise#90620

Related: odoo/upgrade#8121
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>