[FIX] iap_extract: upload batch of documents in postcommit hook
Since commit 7f7ff8d, the upload of a batch of documents isn't safe
against a rollback of the SQL transaction.
If there is a error after some documents were successfully sent to the
OCR server, the transaction will be rollbacked and the IAP credits spent
for those documents will be lost.

This commit attempts at preventing that by moving the upload of the
batch in a postcommit hook, where we can safely commit after each
upload.

opw-4920363

closes odoo/enterprise#89880

Signed-off-by: Ricardo Gomes Rodrigues (rigr) <rigr@odoo.com>