[FIX] account_asset: update only existing accounts
In the post init hook for account_asset we are updating the
chart accounts to add the asset models, and then we load those
assets. The updte should only be performed on accounts that
already exist, otherwise the load will create empty account
records with all values null except the assets, which is not
the intention here. This fix filters the update to existing
accounts only, and filters the asset that use those accounts.

closes odoo/enterprise#90685

X-original-commit: c4bf2de38ff6db176086bb6e5ca187ffe934f9f3
Signed-off-by: Antoine Boonen (aboo) <aboo@odoo.com>