[FIX] appointment: include all users in manual confirm percentage
While the fix has been already implemented in 19.0 in [1], this commit
forwards-port the test of the original commit, as it differs from the
one included (modified) in [1] and treats the case of managing capacity,
and some other use cases not covered by the test in [1].

ORIGINAL ISSUE
==============

Introduced in bce7e94650c337a9046958a7689c81db5b2a4c73,
We now handle capacity when booking users.
They have each a max capacity of user_capacity for each slot.

However, when it comes to the manual confirmation treshold,
we do not account for the total capacity of all users combined,
but consider the maximum to be user_capacity!

This can lead to a lot of 'request' meetings as we reach the
manual confirmation percentage very quickly.

STEPS TO REPRODUCE
==================

1. Create an appointment type based on users, with 'manage capacities'
enabled, 'manual confirmation' enabled and 'when over' 50% total
capacity. Set 2 users and 3 'seats max'. This way, the total capacity
of your appointment is 2 * 3 = 6 per slot.
2. As a public user, take an appointment for 2 capacity in the front
end
3. The meeting will end up as a 'request', even though the capacity
booked is 2/6 < 50%, and should be 'booked'

FIX
===

We now multiply the number of users on the appointment type by the
user_capacity when computing the total capacity and comparing it
to the asked capacity on booking an appointment.

1 : odoo/enterprise@b19e26b0333180ba7a487e14e06385693cf7ab16

Task-4930778

X-original-commit: 73bd9d10008801cc520b8da5f3bd5b0a68dd58d5
Part-of: odoo/enterprise#94446
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Noé Antoine (nan) <nan@odoo.com>