[FIX] l10n_mx_edi_pos: allow refunds with a gift card
**Problem:**
When trying to make a refund in Mexico, with a gift card in the order,
the order should be refunded on said gift card.
But when doing so, two problems arise:
- In the case of a gift card, *line[2].refunded_orderline_id* is
not set, so we are trying to read the property of an undefined
variable.
- Another problem is that the order can not have any line that
have a positive subtotal in the case of a refund. This was not
taking the gift cards into accounts, as it makes sense for them
to be positive in such a case.

**Steps to reproduce:**
- Set your company to a Mexican company
- Make a purchase for a gift card
- Make another purchase for a random product
- Start the refund process and enter the gift card code we created before
- Try to finalize the purchase, a traceback appears
- If we were not trying to read an undefined variable, the purchase would still not go through as some lines are positive in a refund

**Why the fix:**
When refunding with a gift card, we should refill the gift card
by the order's cost instead. With this in mind, it makes sense
for the gift card to have a positive price even though we are in
the context of a refund, which usually only permits negative priced
lines.

opw-4892935

closes odoo/enterprise#91501

X-original-commit: ea78e541ca3a2a83e83a4cce9be6ccaf7c108703
Signed-off-by: Stéphane Vanmeerhaeghe (stva) <stva@odoo.com>
Signed-off-by: Arthur Nanson (artn) <artn@odoo.com>