[FIX] pos_iot: blackbox as scale in PoS fix
If on first boot after flashing the iot box a blackbox is connected the
iot box doesn't have the blackbox driver so it's identified as Adam
scale

With the pos_iot auto configuration we would first send it as a scale to
the database and it would be set as a scale to the selected pos config.

After odoo restart on the iot box it gets detected as blackbox correctly
and its name changes.
Now after it's sent to the db the user ends up with a blackbox in the
scale field in PoS config.

This PR removes the automatic scale configuration so that doesn't
happen

closes odoo/enterprise#89917

X-original-commit: 11c0955ac94bbe08bef660d0538d7cd70d924ee4
Signed-off-by: Louis Travaux (lotr) <lotr@odoo.com>