[IMP] website_appointment: repair inverted toggle for allow_guests
Since 689a2186fbe3b932b90bef3b64716c8443bad551,
The website editor option for allow_guests has an inverted behavior,
being toggled when it should not and vice versa. This is only due to
a small inversion of values when refactoring the website builder code.

Inverse the values when the option is on/off to match expected behavior.

Task-4930982

closes odoo/enterprise#94178

X-original-commit: 6eeee28de7e46e38512e843d9a82d438bde3e7ac
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>