[FIX] whatsapp: fix store dependency issue
`common` folder should be included in all bundles.

This is especially necessary when broadcasting messages on the bus,
because these messages will be received in all tabs, including those
potentially not in the back-end, such the discuss public page, any page
where a live chat is opened, or a portal page with a chatter.

This is also necessary just to have all features properly working in all
context, there is no reason that a whatsapp viewed in the public page
should not have its features working as intended.

runbot-223082

closes odoo/enterprise#89902

X-original-commit: 82a22b2a2745660ba67d26a4e3f1cf80c1de19dd
Signed-off-by: Matthieu Stockbauer (tsm) <tsm@odoo.com>
Signed-off-by: Sébastien Theys (seb) <seb@odoo.com>