[ADD] databases: overview your Odoo databases
The new Databases application allows you to manage a list of other Odoo
databases from a single interface.

Each database is treated as a Project, enabling features such as task
and timesheet management within this context. The list view displays a
variety of KPIs retrieved from each database - for example, the number
of documents in the Inbox folder, the number of draft journal entries,
or the tax returns that need to be submitted.

closes odoo/enterprise#94207

Task-id: 5062431
Related: odoo/odoo#229220
Signed-off-by: Paul Morelle (pmo) <pmo@odoo.com>