[FIX] point_of_sale: appointment kanban filter
Steps to reproduce:

- Open booking in a pos restaurant
- Switch from kanban view to any other view
- The filter on the date is not removed

Issue:
The onRemove method in the kanban_controller is never called.

Fix:
It is not possible to call the code an onWillUmount since onWillUmount
is called after the onMount of the desired view. The deletion of the filter
needs to be handled in the control_panel. Also each time createStartFilter
was called, a new filter was added and never removed. In this commit
if a filter already exists it is simply updated.

Task-4916512

closes odoo/enterprise#90699

X-original-commit: 0ebf4657d7ec874b759edaf13b25041777b1199d
Signed-off-by: David Monnom (moda) <moda@odoo.com>
Signed-off-by: Loukas Wets (lowe) <lowe@odoo.com>