[IMP] appointment{_google_calendar}: use google_meet for online apts
Improve the default value for the event_videocall_source. Instead of
always using Odoo Discuss, use google_meet if the module is installed
and the connector is on (set up and not paused), and only for online
appointments (the location is not set)

The default_get method of appointment.type is overriden to do so. Any
forced value for event_videocall_source will still override this logic.

The method used to get a default video call source in the templates is
removed: all have their value forced to False, except the videocall
template, that will use the new default as we force no value.

Task-4898893

Part-of: odoo/enterprise#88637
Related: odoo/upgrade#8014
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>