[FIX] {l10n_be_}hr_contract_salary: write version fields on version
With the change from the contract to the version model, more adaptation was required for the salary configuration and the offer management.
Before that change, the personal info were always written on the employee model. But now, some of the personal information (like the address or the marital status) need to be written on the new version.

This commit fixes the issue by adding support to directly write the version fields of the salary configurator on the new version.

task-4933745

closes odoo/enterprise#90412

Related: odoo/odoo#219400
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Signed-off-by: Jurgen Gjini (jugj) <jugj@odoo.com>