[FIX] stock_barcode{,_picking_batch}: merge backorder moves on barcode exit
Steps to reproduce:

- Create a storable product tracked by lot
- Create and confirm a receipt for 10 units of that product
- Go to the barcode app and regiter 1 unit
- Exit the picking
- Go back to the picking register the 9 remaining units
- Exit the picking and comeback
> The demand of the receipt has been updated from 10 to 19.

Cause of the issue:

Exiting the picking will launch a call of the `post_barcode_process`
in order to keep track of the changes you made without changing the
initial demand:
https://github.com/odoo/enterprise/blob/c26c21cec14a234021bb13bbc2684334bc64b70b/stock_barcode/models/stock_move.py#L50-L53
THe first time you enter the picking you have a single move with a
quantity of 10. The first time you exit, since you have set a qty_done
of 1 (and hence have updated the quantity of the associated move to 1),
the `split_uncompleted_moves` will then create a move for a quantity of
9 in order to keep 10 units assigned. The second time you enter the
picking both moves are grouped in a single line since the product is
tracked by lot, however, when you update the qty_done of that grouped
line you will actually only update the quantity of the first move line
and its related move. Since the `_truncate_overreserved_moves` was not
designed to handle these grouped lines, it does not notice that the
combined reservation overcomes the actual demand.

opw-4731803

closes odoo/enterprise#92382

X-original-commit: 597eeb627cdb951af0b24ca45c02c4f4d893cea2
Signed-off-by: Steve Van Essche <svs@odoo.com>
Signed-off-by: Lancelot Semal (lase) <lase@odoo.com>