[FIX] project_enterprise: prevent error when rescheduling a task's date
Currently, an error occurs when rescheduling a task's date from the Gantt view.

Steps to Reproduce:

 - Install the project_enterprise module.
 - Go to Settings and enable Task Dependencies.
 - Navigate to Projects, create a project, and add two tasks
   (Task 1 and Task 2) under that project.
 - Open Task 2 > Blocked By, click Add a line, and select Task 1.
 - Return to the project's tasks, switch to Gantt View, and choose
   Auto-Reschedule (Keep Buffer) as the rescheduling method.
 - Double-click on any date in the unassigned block, select Task 1, and drag
   it to next date.

TypeError: '>' not supported between instances of 'bool' and 'datetime.datetime'

This error occurs when rescheduling a task to a forward date, and a dependent
task has no start date. When the parent task is rescheduled, the system
attempts to reschedule the dependent task as well, but since it has no
start date, an error occurs [1] due to condition at [3].

The same error occurs when rescheduling a task to a backward date, and a
depends on task has no deadline date. When the child task is rescheduled,
the system attempts to reschedule the parent task as well, but since it has no
deadline date, an error occurs [2] due to condition at [3].

Steps to Reproduce for backward rescheduling:
 - Follow the first 5 steps from the forward rescheduling steps.
  - Remove the Deadline from Task 1, if it has.
 - Double-click on any date in the unassigned block, select Task 2, and drag
   it to previous date.

This commit ensures that the incorrect condition is removed from [3]. Now,
it checks whether the record is a candidate for moving, and both forward
and backward rescheduling are handled safely.

[1]: https://github.com/odoo/enterprise/blob/aaad26c28fd4bc1e94e99717076ecbcf74f5f357/project_enterprise/models/project_task.py#L1209

[2]: https://github.com/odoo/enterprise/blob/8cc8443dc11c176b55f7eb3efa07db5edb7e6b0c/project_enterprise/models/project_task.py#L1221

[3]: https://github.com/odoo/enterprise/blob/302f1a5e82ffbf383d37f003820fe780654fae2f/project_enterprise/models/project_task.py#L1203

sentry-6738017612

closes odoo/enterprise#90837

X-original-commit: 4b23f8f74b613381cb80eeed44e2e18aa2a6fb79
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Ashutosh Sharma (assh) <assh@odoo.com>