[FIX] account_accountant: fix reconcile of invoice & bill with different partners
Steps to reproduce:
- Install account_accountant

- Create an invoice ($100) for Customer X
- Create a bill ($100) for Vendor Y

- From "Customer Invoices" journal in Accounting dashboard, open "Payments Matching"
- Select the invoice and the bill
- Reconcile them

Issue:
A move is created with 2 lines that will mark both invoice and bill as paid.
However, on both created lines, the partner is "Vendor Y", which is wrong as it will
generate incorrect reports (e.g. Partner Ledger).
One line should have "Vendor Y" as partner and the other line should have "Customer X".
This kind of reconciliation is apparently current in India.

Cause:
When reconciling lines from 2 accounts exactly, a transfer from one account to the
other one is created.
The transfer is made from the account where the absolute amount to move is the lowest.
A transfer is made for each partner that has a move being reconciled with the source
account.
The transfer is crediting (or debiting) the source account and debiting (or crediting)
the reconcile account with the same partner.
This is working most of the time but it totally ignores all the partners linked to
the moves with the reconcile account.

Solution:
Create one line of the transfer with the source account and the partner as before.
For the line on the reconcile account, first check if there's one for the same
partner, currency and sign (of the amount to transfer).
If not, try to find another partner with the same currency and sign.
Otherwise, fall back on the same partner as the line on the source account as it was
done before.

opw-4735147

closes odoo/enterprise#92995

X-original-commit: a8a21a1953fe2345175ed9890be76ea6d40caf82
Related: odoo/odoo#223953
Signed-off-by: Ali Alfie (alal) <alal@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>